Logout component created with wip on redirection when logging out
