+++ web-app/src/pages/settings/index.tsx
@@ -1,3 +1,10 @@
+import Logout from "@/components/home/Logout";
+
 export default function index() {
-  return <div>settings page</div>;
+  return (
+    <>
+      <div>settings page</div>
+      <Logout />
+    </>
+  );
 }
